+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.1.1
+
+- **Add**: `acceptableStatusCodes` property in `DefaultNetworkService`.
+- **Add**: `retry(retryLimit:canRetryClosure:)` to `Observable` extension.
+- **Removed**: `retryWithinErrors` method from `Observable` extension.
+- **Update**: LeadKit to `0.6.x` version
+- **Removed**: `ConnectionError`. (Replaced by `LeadKit.RequestError`)
+- **Removed**: `handleConnectionErrors` from Observable+Extensions
+
+
 ## 0.1.0
 
 - **Add**: support for Swift 3.2 / 4